+++ src/helpers/getAuthorsNames.ts
@@ -1,0 +1,10 @@
+import { mockedAuthorsList } from '../constants';
+
+export function getAuthors(authorsIds: string[]) {
+	return authorsIds
+		.map(
+			(authorId) =>
+				mockedAuthorsList.find((author) => author.id === authorId)?.name
+		)
+		.join(', ');
+}
